+++ example/provider_configuration.tf
@@ -12,12 +12,12 @@ variable "virtual_environment_endpoint" {
   description = "The endpoint for the Proxmox Virtual Environment API (example: https://host:port)"
 }
 
-variable "virtual_environment_username" {
-  type        = "string"
-  description = "The username for the Proxmox Virtual Environment API (example: root@pam)"
-}
-
 variable "virtual_environment_password" {
   type        = "string"
   description = "The password for the Proxmox Virtual Environment API"
+}
+
+variable "virtual_environment_username" {
+  type        = "string"
+  description = "The username and realm for the Proxmox Virtual Environment API (example: root@pam)"
 }