+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 FEATURES:
 
+* **New Data Source:** `proxmox_virtual_environment_access_group`
 * **New Data Source:** `proxmox_virtual_environment_access_groups`
 * **New Data Source:** `proxmox_virtual_environment_version`